requirements.txt
simple_term_menu==1.5.2
requests==2.28.1
python-socketio[client]==5.7.2